restructured code in client into seperate files
